+++ WeaponsManager.cs
@@ -17,7 +17,7 @@ namespace WeaponsManager
     {
         private const string ModId = "com.rsmind.rounds.weaponsmanager";
         private const string ModName = "Weapon Manager";
-        public const string Version = "1.3.4";
+        public const string Version = "1.4.0";
         public const string ModInitials = "WM";
         internal static AssetBundle assets;
         public static WeaponsManager instance { get; private set; }
@@ -37,17 +37,8 @@ namespace WeaponsManager
         void Start()
         {
             instance = this;
-            GameModeManager.AddHook(GameModeHooks.HookGameStart, GameStart);
         }
 
         public static bool Debug = false;
-
-        IEnumerator GameStart(IGameModeHandler gm)
-        {
-            foreach (Player player in PlayerManager.instance.players)
-                player.gameObject.GetOrAddComponent<WeaponManager>();
-            yield break;
-        }
-
     }
 }